celery_app/requirements.txt
fastapi==0.96.0
uvicorn[standard]==0.22.0
pydantic==1.10.9
sqlalchemy==2.0.15
pytz==2024.1
PyJWT==1.7.1
fastapi_jwt_auth==0.5.0
fastapi_versioning==0.10.0
async-timeout==4.0.3
redis==5.0.1
Celery==5.3.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
unoconv
python-dotenv==0.19.1
